lauch Sync Gateway config server from run-test.sh
